refactor baseapp.cpp
