+++ tinyphone-osx/src/Tinyphone-Bridging-Header.h
@@ -2,4 +2,4 @@
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
 
-#include "app.h"
+#include "osxapp.h"
