+++ tinyphone-osx/src/osxapp.h
